Break roles out of main files, adjust search path
Roles are defined at the database cluster level, and so are valid in all
databases in the cluster.  Moving the manipulation of roles into
separate files allows different databases and schemas to run the other
parts of the schema independently.

This facilitates testing, where we might want to use multiple schemas on
the same database to test changes to the schema.  Splitting the schema
up means we don't have to mess with the roles, which must be shared
across all schemas and databases on the cluster.

Setting the search path to just our schema makes it a lot easier to
reason about and modify the schema by reducing redundancy in the DDL.

A test schema can now be created by copying schema.sql and doing a search
and replace of the schema name.  Likewise for obliterate.sql.