Initialize last_updated in schema, grant update to scraper
Setting the default value for the last_updated column to 'epoch' in the
schema ensures that new URLs will be scraped the next time the scraper
runs.  Otherwise, they would not be scraped for the first time until the
calendar gets updated.

The scraper needs to be able to update the value of the last_updated
column, so has been granted UPDATE on this column.